+++ src/strings.cpp
@@ -18,14 +18,14 @@ auto String::Slice() -> ::Slice<u8> {
 auto operator""_s(const char* ptr, unsigned long size) -> String {
     return {
         .data = (u8*)ptr,
-        .size = size,
+        .size = s64(size),
     };
 }
 
 auto As_String(const char* ptr) -> String {
     return {
         .data = (u8*)ptr,
-        .size = strlen(ptr),
+        .size = s64(strlen(ptr)),
     };
 };
 
@@ -37,10 +37,10 @@ auto Destroy_String(Allocator allocator, String& str) -> void {
     };
 }
 
-auto Sprintf(u8* s, usize maxlen, const char* format, ...) -> String {
+auto Sprintf(u8* s, s64 maxlen, const char* format, ...) -> String {
     va_list args;
     va_start(args, format);
-    usize size = vsnprintf((char*)s, maxlen, format, args);
+    s64 size = vsnprintf((char*)s, maxlen, format, args);
     va_end(args);
     return String{
         .data = s,
